Moved expr package out of internal so components using it are importable
